misc: nodejs environment
