+++ lib/java/ydoc-server/src/main/java/org/enso/ydoc/polyfill/nodejs/Encoding.java
@@ -1,4 +1,4 @@
-package org.enso.ydoc.polyfill;
+package org.enso.ydoc.polyfill.nodejs;
 
 import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
