+++ lib/java/ydoc-server/src/main/java/org/enso/ydoc/polyfill/nodejs/Performance.java
@@ -1,4 +1,4 @@
-package org.enso.ydoc.polyfill;
+package org.enso.ydoc.polyfill.nodejs;
 
 import java.util.Arrays;
 import org.enso.ydoc.Polyfill;
