+++ lib/java/ydoc-server/src/main/resources/org/enso/ydoc/polyfill/nodejs/performance.js
